feat: better file organisation feat: some better logs context
